+++ backend/Controllers/MainController.cs
@@ -16,19 +16,19 @@ namespace backend.Controllers
         private readonly DBManager _dbManager;
         private readonly ILogger<MainController> _logger;
         private readonly RecaptchaService _recaptchaService;
-        private readonly EmailService _emailService; 
+        private readonly EmailService _emailService;
 
-    public MainController(
-        ILogger<MainController> logger,
-        RecaptchaService recaptchaService,
-        DBManager dbManager,
-        EmailService emailService)
-    {
-        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-        _recaptchaService = recaptchaService ?? throw new ArgumentNullException(nameof(recaptchaService));
-        _dbManager = dbManager ?? throw new ArgumentNullException(nameof(dbManager));
-        _emailService = emailService ?? throw new ArgumentNullException(nameof(emailService));
-    }
+        public MainController(
+            ILogger<MainController> logger,
+            RecaptchaService recaptchaService,
+            DBManager dbManager,
+            EmailService emailService)
+        {
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _recaptchaService = recaptchaService ?? throw new ArgumentNullException(nameof(recaptchaService));
+            _dbManager = dbManager ?? throw new ArgumentNullException(nameof(dbManager));
+            _emailService = emailService ?? throw new ArgumentNullException(nameof(emailService));
+        }
 
 
         [HttpPost("registration")]
@@ -74,35 +74,35 @@ namespace backend.Controllers
                         )";
 
                     if (_dbManager.insert(dbInsert, query))
+                    {
+                        _logger.LogInformation($"User {registration.Email} registered successfully.");
+                        _logger.LogInformation($"First Name: {registration.FirstName}");
+                        _logger.LogInformation($"Last Name: {registration.LastName}");
+                        _logger.LogInformation($"Email: {registration.Email}");
+                        _logger.LogInformation($"Role: {registration.Role}");
+                        _logger.LogInformation($"Token: {token}");
+
+                        // ✅ Send the verification email
+                        string verificationUrl = $"http://localhost:5172/api/main/confirm?token={token}";
+                        string emailBody = $@"
+                            <p>Click the link below to verify your account:</p>
+                            <a href=""{verificationUrl}"">Verify My Account</a>
+                        ";
+
+                        var emailSent = await _emailService.SendEmail(registration.Email, "Verify Your Account", emailBody);
+                        if (!emailSent)
                         {
-                            _logger.LogInformation($"User {registration.Email} registered successfully.");
-                            _logger.LogInformation($"First Name: {registration.FirstName}");
-                            _logger.LogInformation($"Last Name: {registration.LastName}");
-                            _logger.LogInformation($"Email: {registration.Email}");
-                            _logger.LogInformation($"Role: {registration.Role}");
-                            _logger.LogInformation($"Token: {token}");
-
-                            string verificationUrl = $"http://localhost:5172/api/verify/confirm?token={token}";
-                            string emailBody = $@"
-                                <p>Click the link below to verify your account:</p>
-                                <a href=""{verificationUrl}"">Verify My Account</a>
-                            ";
-
-                            var emailSent = await _emailService.SendEmail(registration.Email, "Verify Your Account", emailBody);
-                            if (!emailSent)
-                            {
-                                _logger.LogError("Failed to send verification email.");
-                                return StatusCode(500, new { message = "Failed to send verification email." });
-                            }
-
-                            return Ok(new { message = "User registered successfully. Verification email sent." });
-                        }
-                        else
-                        {
-                            _logger.LogError("Failed to register user; database error.");
-                            return BadRequest(new { message = "Failed to register user; database error." });
+                            _logger.LogError("Failed to send verification email.");
+                            return StatusCode(500, new { message = "Failed to send verification email." });
                         }
 
+                        return Ok(new { message = "User registered successfully. Verification email sent." });
+                    }
+                    else
+                    {
+                        _logger.LogError("Failed to register user; database error.");
+                        return BadRequest(new { message = "Failed to register user; database error." });
+                    }
                 }
             }
             catch (Exception ex)
@@ -111,6 +111,107 @@ namespace backend.Controllers
                 return StatusCode(StatusCodes.Status500InternalServerError, new { message = "Failed to register user." });
             }
         }
+
+
+        [HttpPost("send-code")]
+        public async Task<IActionResult> SendVerificationCode([FromBody] EmailRequest request)
+        {
+            using var conn = _dbManager.connect();
+
+            string selectSql = $@"
+                SELECT verified
+                FROM accounts
+                WHERE email = '{request.Email}'
+                LIMIT 1
+            ";
+
+            var rows = _dbManager.select(conn, selectSql);
+
+            if (rows.Count == 0)
+            {
+                _dbManager.close(conn);
+                return BadRequest(new { message = "Account does not exist" });
+            }
+
+            bool isVerified = (bool)rows[0]["verified"];
+            if (isVerified)
+            {
+                _dbManager.close(conn);
+                return Ok(new { message = "Account is already verified" });
+            }
+
+            bool success = await _emailService.SendVerificationCode(request.Email);
+
+            _dbManager.close(conn);
+
+            if (success)
+            {
+                return Ok(new { message = "Verification code sent successfully" });
+            }
+            else
+            {
+                return StatusCode(500, new { message = "Failed to send verification code" });
+            }
+        }
+
+
+        [HttpGet("confirm")]
+        public IActionResult ConfirmAccount(string token)
+        {
+            using var conn = _dbManager.connect();
+
+            try
+            {
+                string selectSql = $@"
+                    SELECT email, verified
+                    FROM accounts
+                    WHERE verification_token = '{token}'
+                    LIMIT 1
+                ";
+
+                var rows = _dbManager.select(conn, selectSql);
+
+                if (rows.Count == 0)
+                {
+                    return BadRequest(new { message = "Invalid or expired token." });
+                }
+
+                bool isVerified = (bool)rows[0]["verified"];
+                if (isVerified)
+                {
+                    return Ok(new { message = "Account is already verified." });
+                }
+
+                string updateSql = $@"
+                    UPDATE accounts
+                    SET verified = TRUE,
+                        verification_token = NULL
+                    WHERE verification_token = '{token}'
+                ";
+
+                int rowsAffected = _dbManager.update(conn, updateSql);
+                conn.Close(); 
+
+                if (rowsAffected > 0)
+                {
+                    return Ok(new { message = "Account verified successfully." });
+                }
+                else
+                {
+                    return BadRequest(new { message = "Failed to verify account." });
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"[ERROR] Failed to verify account: {ex.Message}");
+                return StatusCode(500, new { message = "Internal server error." });
+            }
+            finally
+            {
+                conn.Close();
+            }
+        }
+
 
         [HttpPost("login")]
         public async Task<IActionResult> Login([FromBody] LoginModel loginCredentials)
