--- backend/Models/EmailRequestModel.cs
@@ -1,7 +1,0 @@
-namespace backend.Models
-{
-    public class EmailRequest
-    {
-        public string Email { get; set; }
-    }
-}
