--- backend/Models/EmilsettingsModel.cs
@@ -1,11 +1,0 @@
-namespace backend.Models
-{
-    public class EmailSettings
-    {
-        public string SmtpServer { get; set; }
-        public int Port { get; set; }
-        public string Username { get; set; }
-        public string Password { get; set; }
-        public bool EnableSsl { get; set; } 
-    }
-}
