+++ backend/Models/MainModel.cs
@@ -21,4 +21,24 @@ namespace backend.Models
         public required string Password { get; set; }
         public required string RecaptchaToken { get; set; }
     }
+
+    public class VerificationRequest
+    {
+        public string Email { get; set; }
+        public string Code { get; set; }
+    }
+
+    public class EmailSettings
+    {
+        public string SmtpServer { get; set; }
+        public int Port { get; set; }
+        public string Username { get; set; }
+        public string Password { get; set; }
+        public bool EnableSsl { get; set; } 
+    }
+
+    public class EmailRequest
+    {
+        public string Email { get; set; }
+    }
 }
