--- backend/Models/VerificationRequestModel.cs
@@ -1,8 +1,0 @@
-namespace backend.Models
-{
-    public class VerificationRequest
-    {
-        public string Email { get; set; }
-        public string Code { get; set; }
-    }
-}
